requirements.txt
flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
waitress==3.0.0
lxml==5.1.0